harden durable provenance ingest
