+++ crates/gents/src/hook/persistence/approval.rs
@@ -26,6 +26,7 @@ const APPROVAL_POLL_INTERVAL: Duration = Duration::from_millis(750);
 
 #[derive(Debug, Deserialize)]
 struct ApprovalRow {
+    tool_call_doc_id: Option<String>,
     decision: Option<String>,
     reason: Option<String>,
     approver_did: Option<String>,
@@ -37,28 +38,32 @@ enum ApprovalDecision {
     Denied { reason: Option<String> },
 }
 
+struct ApprovalPoll {
+    decision: Option<ApprovalDecision>,
+    ignored_legacy_unbound: bool,
+}
+
 async fn first_approval_decision(
     node: &defra_node::EmbeddedNode,
     agent_did: &str,
     tool_call_doc_id: &str,
     tool_call_id: &str,
-) -> anyhow::Result<Option<ApprovalDecision>> {
+) -> anyhow::Result<ApprovalPoll> {
     // Physical tool-call identity is the authorization boundary. A legacy
     // logical-only approval must not authorize a different document that
     // happens to reuse the same tool_call_id.
-    let escaped_tool_call_doc_id = escape_graphql_string(tool_call_doc_id);
     let escaped_tool_call_id = escape_graphql_string(tool_call_id);
     let escaped_agent_did = escape_graphql_string(agent_did);
     let query = format!(
         r#"{{
             AgentToolApproval(
                 filter: {{
-                    tool_call_doc_id: {{ _eq: "{escaped_tool_call_doc_id}" }},
                     tool_call_id: {{ _eq: "{escaped_tool_call_id}" }},
                     agent_did: {{ _eq: "{escaped_agent_did}" }}
                 }},
                 order: {{ created_at: ASC }}
             ) {{
+                tool_call_doc_id
                 decision
                 reason
                 approver_did
@@ -80,7 +85,17 @@ async fn first_approval_decision(
         .context("decode AgentToolApproval rows")?
         .unwrap_or_default();
 
+    let saw_legacy_unbound = rows.iter().any(|row| {
+        row.tool_call_doc_id
+            .as_deref()
+            .map(str::trim)
+            .map(str::is_empty)
+            .unwrap_or(true)
+    });
     for row in rows {
+        if row.tool_call_doc_id.as_deref().map(str::trim) != Some(tool_call_doc_id) {
+            continue;
+        }
         match row.decision.as_deref() {
             Some("approved") => {
                 tracing::info!(
@@ -88,7 +103,10 @@ async fn first_approval_decision(
                     approver_did = row.approver_did.as_deref().unwrap_or(""),
                     "tool call approved"
                 );
-                return Ok(Some(ApprovalDecision::Approved));
+                return Ok(ApprovalPoll {
+                    decision: Some(ApprovalDecision::Approved),
+                    ignored_legacy_unbound: saw_legacy_unbound,
+                });
             }
             Some("denied") => {
                 tracing::info!(
@@ -96,7 +114,10 @@ async fn first_approval_decision(
                     approver_did = row.approver_did.as_deref().unwrap_or(""),
                     "tool call denied"
                 );
-                return Ok(Some(ApprovalDecision::Denied { reason: row.reason }));
+                return Ok(ApprovalPoll {
+                    decision: Some(ApprovalDecision::Denied { reason: row.reason }),
+                    ignored_legacy_unbound: saw_legacy_unbound,
+                });
             }
             other => {
                 tracing::warn!(
@@ -107,7 +128,10 @@ async fn first_approval_decision(
             }
         }
     }
-    Ok(None)
+    Ok(ApprovalPoll {
+        decision: None,
+        ignored_legacy_unbound: saw_legacy_unbound,
+    })
 }
 
 impl DefraSessionHook {
@@ -116,6 +140,7 @@ impl DefraSessionHook {
         tool_name: &str,
         internal_call_id: &str,
     ) -> anyhow::Result<ToolCallHookAction> {
+        let mut warned_legacy_unbound = false;
         loop {
             let observed = {
                 let map = self.in_flight_lifecycles.lock().await;
@@ -180,14 +205,22 @@ impl DefraSessionHook {
                 )));
             }
 
-            match first_approval_decision(
+            let poll = first_approval_decision(
                 &self.node,
                 &self.agent_did,
                 &tool_call_doc_id,
                 internal_call_id,
             )
-            .await?
-            {
+            .await?;
+            if poll.ignored_legacy_unbound && !warned_legacy_unbound {
+                warned_legacy_unbound = true;
+                tracing::warn!(
+                    tool_call_id = internal_call_id,
+                    expected_tool_call_doc_id = %tool_call_doc_id,
+                    "ignoring legacy AgentToolApproval without physical tool-call provenance",
+                );
+            }
+            match poll.decision {
                 Some(ApprovalDecision::Approved) => {
                     let approved = {
                         let mut map = self.in_flight_lifecycles.lock().await;
@@ -266,19 +299,17 @@ mod tests {
             response.errors
         );
 
-        assert!(first_approval_decision(
-            &node,
-            "did:test:agent",
-            "different-tool-doc",
-            "call-shared",
-        )
-        .await
-        .unwrap()
-        .is_none());
+        let legacy_poll =
+            first_approval_decision(&node, "did:test:agent", "different-tool-doc", "call-shared")
+                .await
+                .unwrap();
+        assert!(legacy_poll.decision.is_none());
+        assert!(legacy_poll.ignored_legacy_unbound);
         assert!(matches!(
             first_approval_decision(&node, "did:test:agent", "tool-doc-exact", "call-shared",)
                 .await
-                .unwrap(),
+                .unwrap()
+                .decision,
             Some(ApprovalDecision::Denied { .. })
         ));
         node.shutdown().await;
