+++ crates/gents/src/run_timeline_fetch/context_loaders.rs
@@ -1,0 +1,65 @@
+use super::*;
+
+pub(super) async fn load_timeline_session(
+    access: &ConfigAccess,
+    session_id: &str,
+) -> Result<Option<TimelineSessionRow>> {
+    let query = format!(
+        r#"{{
+            AgentSession(
+                filter: {{ session_id: {{ _eq: "{}" }} }},
+                limit: 1
+            ) {{
+                _docID
+                session_id
+                agent_name
+                behavior_id
+                started
+                ended
+                status
+            }}
+        }}"#,
+        escape_graphql_string(session_id)
+    );
+    Ok(
+        load_rows::<TimelineSessionRow>(access, "AgentSession", &query)
+            .await?
+            .into_iter()
+            .next(),
+    )
+}
+
+pub(super) async fn load_timeline_conversation(
+    access: &ConfigAccess,
+    session_id: &str,
+) -> Result<Option<TimelineConversationRow>> {
+    let query = format!(
+        r#"{{
+            AgentConversation(
+                filter: {{ session_id: {{ _eq: "{}" }} }},
+                limit: 1
+            ) {{
+                _docID
+                session_id
+                agent_name
+                agent_did
+                behavior_id
+                title
+                title_source
+                preview_text
+                status
+                created_at
+                updated_at
+                latest_request_id
+                forked_from_session_id
+            }}
+        }}"#,
+        escape_graphql_string(session_id)
+    );
+    Ok(
+        load_rows::<TimelineConversationRow>(access, "AgentConversation", &query)
+            .await?
+            .into_iter()
+            .next(),
+    )
+}
