+++ crates/gents/src/streaming.rs
@@ -160,6 +160,15 @@ impl DefraStreamWriter {
         behavior_id: &str,
         requester_did: Option<&str>,
     ) -> Result<String> {
+        let resolved_request_doc_id;
+        let request_doc_id = match request_doc_id.map(str::trim).filter(|id| !id.is_empty()) {
+            Some(request_doc_id) => Some(request_doc_id),
+            None => {
+                resolved_request_doc_id =
+                    crate::request_binding::require_request_doc_id(&self.node, request_id).await?;
+                Some(resolved_request_doc_id.as_str())
+            }
+        };
         self.begin_inner(
             session_id,
             request_id,
@@ -593,7 +602,7 @@ impl DefraStreamWriter {
 
 impl StreamWriter for DefraStreamWriter {
     async fn begin(&self, session_id: &str, request_id: &str, behavior_id: &str) -> Result<String> {
-        self.begin_inner(session_id, request_id, None, behavior_id, None)
+        self.begin_with_requester_did(session_id, request_id, None, behavior_id, None)
             .await
     }
