+++ crates/gents/src/watcher/tests.rs
@@ -77,6 +77,38 @@ fn validate_accepts_background_completion_lineage_without_tool_call_link() {
 }
 
 #[test]
+fn validate_accepts_depth_zero_background_completion_control_lineage() {
+    let req = AgentRequest {
+        subagent_depth: 0,
+        metadata: Some(
+            r#"{"queue":{"source":"background_completion","policy":"coalesce","key":"background_completion:session-1","queued_after_request_id":"goal-1"},"background_completion_wake_version":1}"#
+                .to_string(),
+        ),
+        caused_by_parent_request_id: Some("goal-parent-1".to_string()),
+        caused_by_parent_request_doc_id: Some("goal-parent-doc-1".to_string()),
+        caused_by_parent_tool_call_id: None,
+        caused_by_parent_tool_call_doc_id: None,
+        ..base_request()
+    };
+    assert!(validate_agent_request_subagent_coherence(&req).is_ok());
+}
+
+#[test]
+fn validate_accepts_depth_zero_steering_control_lineage() {
+    let req = AgentRequest {
+        subagent_depth: 0,
+        metadata: Some(
+            r#"{"queue":{"source":"steering","policy":"append","key":null,"queued_after_request_id":null}}"#
+                .to_string(),
+        ),
+        caused_by_parent_request_id: Some("parent-req-1".to_string()),
+        caused_by_parent_request_doc_id: Some("parent-req-doc-1".to_string()),
+        ..base_request()
+    };
+    assert!(validate_agent_request_subagent_coherence(&req).is_ok());
+}
+
+#[test]
 fn validate_rejects_mixed_parent_linkage_tool_call_id_only() {
     let req = AgentRequest {
         subagent_depth: 1,
@@ -673,6 +705,12 @@ async fn pending_requests_quarantines_incoherent_row_without_hiding_valid_work()
         vec!["req-valid-pending"],
         "an incoherent row must not poison or queue-block coherent work"
     );
+    let terminal = request_terminal_fields(node.as_ref(), "req-incoherent-pending").await;
+    assert_eq!(terminal["status"], "failed");
+    assert_eq!(terminal["lifecycle_state"], "failed");
+    assert!(terminal["failure_reason"]
+        .as_str()
+        .is_some_and(|reason| reason.contains("incoherent durable lineage")));
 }
 
 #[tokio::test]
@@ -700,7 +738,7 @@ async fn incoherent_live_row_still_blocks_a_second_same_session_claim() {
 }
 
 #[tokio::test]
-async fn try_fetch_request_rejects_incoherent_subagent_linkage() {
+async fn try_fetch_request_terminalizes_incoherent_subagent_linkage() {
     let node = test_node().await;
     crate::ensure_runtime_schemas(node.as_ref()).await.unwrap();
 
@@ -709,10 +747,9 @@ async fn try_fetch_request_rejects_incoherent_subagent_linkage() {
         insert_incoherent_agent_request(node.as_ref(), agent_did, "req-incoherent-fetch").await;
 
     let watcher = DefraWatcher::new(node.clone(), agent_did);
-    let result = watcher.try_fetch_request(&doc_id).await;
-    assert!(
-        result.is_err(),
-        "try_fetch_request must fail for incoherent subagent linkage, got: {:?}",
-        result
-    );
+    let result = watcher.try_fetch_request(&doc_id).await.unwrap();
+    assert!(result.is_none());
+    let terminal = request_terminal_fields(node.as_ref(), "req-incoherent-fetch").await;
+    assert_eq!(terminal["status"], "failed");
+    assert_eq!(terminal["lifecycle_state"], "failed");
 }
